refactored toolbelt for edition command
