+++ src/api/modules/utils.ts
@@ -12,7 +12,7 @@ import { ManifestEditor } from '../manifest'
 import { getAppRoot } from '../manifest/ManifestUtil'
 import { SessionManager } from '../session/SessionManager'
 import log from '../logger'
-import { promptConfirm } from '../../modules/prompts'
+import { promptConfirm } from './prompts'
 import { execSync } from 'child-process-es6-promise'
 import { diffArrays } from 'diff'
 import { existsSync } from 'fs-extra'
