+++ src/modules/apps/link.ts
@@ -23,7 +23,7 @@ import { randomBytes } from 'crypto'
 import { readFileSync } from 'fs'
 import { SessionManager } from '../../api/session/SessionManager'
 import { YarnFilesManager } from '../../api/files/YarnFilesManager'
-import authLogin, { LoginOptions } from '../auth/login'
+import authLogin, { LoginOptions } from '../../api/modules/auth/login'
 import chalk from 'chalk'
 import chokidar from 'chokidar'
 import debounce from 'debounce'
