+++ src/modules/apps/publish.ts
@@ -9,9 +9,9 @@ import { getAppRoot } from '../../api/manifest/ManifestUtil'
 import { toAppLocator } from '../../api/locator'
 import log from '../../api/logger'
 
-import { switchAccount, returnToPreviousAccount } from '../auth/switch'
+import { switchAccount, returnToPreviousAccount } from '../../api/modules/auth/switch'
 import { listenBuild } from '../../api/modules/build'
-import { promptConfirm } from '../prompts'
+import { promptConfirm } from '../../api/modules/prompts'
 import { runYarnIfPathExists } from '../utils'
 import { listLocalFiles } from '../../api/modules/apps/file'
 import { ProjectUploader } from '../../api/modules/apps/ProjectUploader'
