+++ src/modules/workspace/abtest/finish.ts
@@ -3,7 +3,7 @@ import enquirer from 'enquirer'
 import { map, prop } from 'ramda'
 
 import log from '../../../api/logger'
-import { promptConfirm } from '../../prompts'
+import { promptConfirm } from '../../../api/modules/prompts'
 import { default as abTestStatus } from './status'
 import { abtester, installedABTester } from './utils'
 import { SessionManager } from '../../../api/session/SessionManager'
