+++ src/modules/workspace/common/edition.ts
@@ -3,7 +3,7 @@ import { Sponsor } from '../../../api/clients/IOClients/apps/Sponsor'
 import { ErrorKinds } from '../../../api/error/ErrorKinds'
 import { ErrorReport } from '../../../api/error/ErrorReport'
 import log from '../../../api/logger'
-import { promptConfirm } from '../../prompts'
+import { promptConfirm } from '../../../api/modules/prompts'
 import setEditionCmd from '../../sponsor/setEdition'
 
 const recommendedEdition = 'vtex.edition-store@2.x'
