+++ src/modules/workspace/promote.ts
@@ -3,7 +3,7 @@ import { CommandError } from '../../api/error/errors'
 import { createWorkspacesClient } from '../../api/clients/IOClients/infra/Workspaces'
 import { SessionManager } from '../../api/session/SessionManager'
 import log from '../../api/logger'
-import { promptConfirm } from '../prompts'
+import { promptConfirm } from '../../api/modules/prompts'
 import useCmd from './use'
 
 const { promote, get } = createWorkspacesClient()
