+++ src/oclif/hooks/init.ts
@@ -14,7 +14,7 @@ import * as pkg from '../../../package.json'
 import * as conf from '../../conf'
 import { checkAndOpenNPSLink } from '../../nps'
 import { Metric } from '../../api/metrics/MetricReport'
-import authLogin from '../../modules/auth/login'
+import authLogin from '../../api/modules/auth/login'
 import { CommandError, SSEConnectionError } from '../../api/error/errors'
 import { MetricNames } from '../../api/metrics/MetricNames'
 import { SessionManager } from '../../api/session/SessionManager'
